+++ Runtime/Scripts/AudioStreamBase.cs
@@ -1,9 +1,0 @@
-namespace Unity.RenderStreaming
-{
-    /// <summary>
-    /// 
-    /// </summary>
-    public abstract class AudioStreamBase : StreamSourceBase
-    {
-    }
-}