+++ Runtime/Scripts/RenderStreamingInternal.cs
@@ -98,7 +98,7 @@ namespace Unity.RenderStreaming
         private readonly Dictionary<string, PeerConnection> _mapConnectionIdAndPeer =
             new Dictionary<string, PeerConnection>();
         private bool _runningResendCoroutine;
-        private float _resendInterval = 1.0f;
+        private float _resendInterval = 3.0f;
 
         static List<RenderStreamingInternal> s_list = new List<RenderStreamingInternal>();
 
@@ -221,7 +221,7 @@ namespace Unity.RenderStreaming
         /// </summary>
         /// <param name="connectionId"></param>
         /// <param name="track"></param>
-        public RTCRtpTransceiver AddTrack(string connectionId, MediaStreamTrack track)
+        public RTCRtpTransceiver AddSenderTrack(string connectionId, MediaStreamTrack track)
         {
             var peer = _mapConnectionIdAndPeer[connectionId];
             RTCRtpSender sender = peer.peer.AddTrack(track);
@@ -232,6 +232,17 @@ namespace Unity.RenderStreaming
             // Please remove the line after supporting the hardware decoder.
             transceiver.Direction = RTCRtpTransceiverDirection.SendOnly;
             return transceiver;
+        }
+
+        /// <summary>
+        ///
+        /// </summary>
+        /// <param name="connectionId"></param>
+        /// <param name="track"></param>
+        public void RemoveSenderTrack(string connectionId, MediaStreamTrack track)
+        {
+            var sender = GetSenders(connectionId).First(s => s.Track == track);
+            _mapConnectionIdAndPeer[connectionId].peer.RemoveTrack(sender);
         }
 
         /// <summary>
@@ -266,22 +277,13 @@ namespace Unity.RenderStreaming
         ///
         /// </summary>
         /// <param name="connectionId"></param>
-        /// <param name="track"></param>
-        public void RemoveTrack(string connectionId, MediaStreamTrack track)
-        {
-            var sender = GetSenders(connectionId).First(s => s.Track == track);
-            _mapConnectionIdAndPeer[connectionId].peer.RemoveTrack(sender);
-        }
-
-        /// <summary>
-        ///
-        /// </summary>
-        /// <param name="connectionId"></param>
         /// <param name="name"></param>
         /// <returns></returns>
         public RTCDataChannel CreateChannel(string connectionId, string name)
         {
             RTCDataChannelInit conf = new RTCDataChannelInit();
+            if (string.IsNullOrEmpty(name))
+                name = Guid.NewGuid().ToString();
             return _mapConnectionIdAndPeer[connectionId].peer.CreateDataChannel(name, conf);
         }
 
@@ -311,9 +313,21 @@ namespace Unity.RenderStreaming
         ///
         /// </summary>
         /// <param name="connectionId"></param>
+        /// <param name="track"></param>
+        /// <returns></returns>
+        public IEnumerable<RTCRtpTransceiver> GetTransceivers(string connectionId)
+        {
+            return _mapConnectionIdAndPeer[connectionId].peer.GetTransceivers();
+        }
+
+        /// <summary>
+        ///
+        /// </summary>
+        /// <param name="connectionId"></param>
         public void SendOffer(string connectionId)
         {
-            var pc = _mapConnectionIdAndPeer[connectionId];
+            if (!_mapConnectionIdAndPeer.TryGetValue(connectionId, out var pc))
+                return;
             if (!IsStable(connectionId))
             {
                 if (!pc.waitingAnswer)
@@ -325,7 +339,6 @@ namespace Unity.RenderStreaming
                 _signaling.SendOffer(connectionId, pc.peer.LocalDescription);
                 return;
             }
-
             _startCoroutine(SendOfferCoroutine(connectionId, pc));
         }
 
@@ -344,10 +357,15 @@ namespace Unity.RenderStreaming
             {
                 foreach (var pair in _mapConnectionIdAndPeer.Where(x => x.Value.waitingAnswer))
                 {
-                    _signaling.SendOffer(pair.Key, pair.Value.peer.LocalDescription);
-                }
+                    float timeout = pair.Value.timeSinceStartWaitingAnswer + _resendInterval;
 
-                yield return new WaitForSeconds(_resendInterval);
+                    if (timeout < Time.realtimeSinceStartup)
+                    {
+                        _signaling.SendOffer(pair.Key, pair.Value.peer.LocalDescription);
+                        pair.Value.RestartTimerForWaitingAnswer();
+                    }
+                }
+                yield return 0;
             }
         }
 
@@ -390,8 +408,7 @@ namespace Unity.RenderStreaming
             {
                 _signaling.SendCandidate(connectionId, candidate);
             };
-            pc.OnIceConnectionChange = state => OnIceConnectionChange(connectionId, state);
-
+            pc.OnConnectionStateChange = state => OnConnectionStateChange(connectionId, state);
             pc.OnTrack = trackEvent =>
             {
                 onAddReceiver?.Invoke(connectionId, trackEvent.Receiver);
@@ -416,14 +433,14 @@ namespace Unity.RenderStreaming
             onAddChannel?.Invoke(connectionId, channel);
         }
 
-        void OnIceConnectionChange(string connectionId, RTCIceConnectionState state)
+        void OnConnectionStateChange(string connectionId, RTCPeerConnectionState state)
         {
             switch (state)
             {
-                case RTCIceConnectionState.Connected:
+                case RTCPeerConnectionState.Connected:
                     onConnect?.Invoke(connectionId);
                     break;
-                case RTCIceConnectionState.Disconnected:
+                case RTCPeerConnectionState.Disconnected:
                     onDisconnect?.Invoke(connectionId);
                     break;
             }
@@ -439,6 +456,9 @@ namespace Unity.RenderStreaming
         {
             // waiting other setLocalDescription process
             yield return new WaitWhile(() => !IsStable(connectionId));
+
+            if (!ExistConnection(connectionId))
+                yield break;
 
             Assert.AreEqual(pc.peer.SignalingState, RTCSignalingState.Stable,
                 $"{pc} negotiationneeded always fires in stable state");