+++ Samples~/Example/Broadcast/BroadcastSample.cs
@@ -1,10 +1,24 @@
 using UnityEngine;
+using UnityEngine.InputSystem.Layouts;
+using UnityEngine.InputSystem.XR;
 
 namespace Unity.RenderStreaming.Samples
 {
+    using InputSystem = UnityEngine.InputSystem.InputSystem;
+
     class BroadcastSample : MonoBehaviour
     {
         [SerializeField] RenderStreaming renderStreaming;
+
+        private void Awake()
+        {
+#if URS_USE_AR_FOUNDATION
+            InputSystem.RegisterLayout<UnityEngine.XR.ARSubsystems.HandheldARInputDevice>(
+                matches: new InputDeviceMatcher()
+                    .WithInterface(XRUtilities.InterfaceMatchAnyVersion)
+                );
+#endif
+        }
 
         // Start is called before the first frame update
         void Start()
@@ -14,7 +28,6 @@ namespace Unity.RenderStreaming.Samples
             renderStreaming.Run(
                 hardwareEncoder: RenderStreamingSettings.EnableHWCodec,
                 signaling: RenderStreamingSettings.Signaling);
-
         }
     }
 }