+++ Samples~/Example/Gyro/GyroSample.cs
@@ -6,13 +6,15 @@ using Gyroscope = UnityEngine.InputSystem.Gyroscope;
 
 namespace Unity.RenderStreaming.Samples
 {
+    using InputSystem = UnityEngine.InputSystem.InputSystem;
+
     class GyroSample : MonoBehaviour
     {
 #pragma warning disable 0649
             [SerializeField] private RenderStreaming renderStreaming;
             [SerializeField] private Button sendOfferButton;
             [SerializeField] private RawImage remoteVideoImage;
-            [SerializeField] private ReceiveVideoViewer receiveVideoViewer;
+            [SerializeField] private VideoStreamReceiver receiveVideoViewer;
             [SerializeField] private SingleConnection connection;
             [SerializeField] private Text textVelocityX;
             [SerializeField] private Text textVelocityY;